+++ sg_compute_specs/open_design/service/Open_Design__Service.py
@@ -32,8 +32,8 @@ class Open_Design__Service(Type_Safe):
         from sg_compute_specs.open_design.service.Open_Design__AWS__Client       import Open_Design__AWS__Client
         from sg_compute_specs.open_design.service.Open_Design__User_Data__Builder import Open_Design__User_Data__Builder
         from sg_compute_specs.open_design.service.Open_Design__Stack__Mapper     import Open_Design__Stack__Mapper
-        from sg_compute.helpers.networking.Caller__IP__Detector                   import Caller__IP__Detector
-        from sg_compute.helpers.networking.Stack__Name__Generator                 import Stack__Name__Generator
+        from sg_compute.platforms.ec2.networking.Caller__IP__Detector                   import Caller__IP__Detector
+        from sg_compute.platforms.ec2.networking.Stack__Name__Generator                 import Stack__Name__Generator
         self.aws_client        = Open_Design__AWS__Client       ().setup()
         self.user_data_builder = Open_Design__User_Data__Builder()
         self.mapper            = Open_Design__Stack__Mapper     ()